Move what is now an |impl Element| block to element.rs.
This is a simple cut/paste.
